buildx not necessary (see details below)
Without buildx you need to seriously watch your cache. Ultimately
the solution here still as a hole, because the FROM image(s) in
the Dockerfile will end up cached in whatever architecture
goes last. We are relying on the build server not caching things,
but be wary when doing local multi-architecture builds